refactor(frontend): split provider config utilities
